requirements.txt
ollama
rich
langchain
langchain-community
langchain-postgres
psycopg[binary]
pypdf
python-dotenv
fastapi
uvicorn
sqlalchemy
pydantic
